requirements.txt
torch
opencv-python
numpy
python-Levenshtein
ultralytics
python-bidi
arabic-reshaper
Pillow
